+++ frontend/src/components/Cam/Nickname/NickNameForm.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import { UserInfo } from '../../types/cam';
+import { UserInfo } from '../../../types/cam';
 
 const Container = styled.div`
   width: 100vw;
