+++ frontend/src/components/Cam/Screen/DefaultScreen.tsx
