+++ frontend/src/components/Cam/UserList/UserListTab.tsx
@@ -1,12 +1,12 @@
 import React, { useContext } from 'react';
 import styled from 'styled-components';
 
-import UserScreen from './UserScreen';
-import { CamStoreContext } from './CamStore';
-import LocalUserScreen from './LocalUserScreen';
-import Draggable from '../core/Draggable';
-import type { Screen } from '../../types/cam';
-import { ToggleStoreContext } from './ToggleStore';
+import UserScreen from '../Screen/UserScreen';
+import { CamStoreContext } from '../CamStore';
+import LocalUserScreen from '../Screen/LocalUserScreen';
+import Draggable from '../../core/Draggable';
+import type { Screen } from '../../../types/cam';
+import { ToggleStoreContext } from '../ToggleStore';
 
 const Container = styled.div<{ isActive: boolean }>`
   position: absolute;
